// src/app/shared/models/notice/notice.model.ts
export class NoticeModel {
    title: string;
    content: string;
    status: string;
    publishTime: string;
}

export class UpdataModel {
    id: number;
    title: string;
    content: string;
    status: string;
    publishTime: string;
}

export class StatusModel {
    id: number;
    status: string;
}